requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
motor==3.3.2
pymongo==4.6.1
aiohttp==3.9.3
aiofiles==23.2.1
dnspython==2.5.0
python-dotenv==1.0.1
certifi==2024.2.2
